+++ packages/webpack-cli/lib/webpack-cli.js
@@ -1,11 +1,15 @@
+const { packageExists } = require('./utils/package-exists');
+const webpack = packageExists('webpack') ? require('webpack') : undefined;
+const logger = require('./utils/logger');
 const webpackMerge = require('webpack-merge');
 const GroupHelper = require('./utils/GroupHelper');
-const { Compiler } = require('./utils/Compiler');
 const { groups, core } = require('./utils/cli-flags');
 const argParser = require('./utils/arg-parser');
 const { outputStrategy } = require('./utils/merge-strategies');
 const { toKebabCase } = require('./utils/helpers');
 const assignFlagDefaults = require('./utils/flag-defaults');
+const { writeFileSync } = require('fs');
+const { options: coloretteOptions } = require('colorette');
 
 // CLI arg resolvers
 const handleConfigResolution = require('./groups/ConfigGroup');
@@ -19,7 +23,6 @@ class WebpackCLI extends GroupHelper {
     constructor() {
         super();
         this.groupMap = new Map();
-        this.compilation = new Compiler();
         this.compilerConfiguration = {};
         this.outputConfiguration = {};
     }
@@ -205,24 +208,173 @@ class WebpackCLI extends GroupHelper {
     async processArgs(args, cliOptions) {
         this.setMappedGroups(args, cliOptions);
         this.resolveGroups(args);
-        const groupResult = await this.runOptionGroups(args);
-        return groupResult;
+
+        return this.runOptionGroups(args);
+    }
+
+    createCompiler(options) {
+        let compiler;
+
+        try {
+            compiler = webpack(options);
+        } catch (error) {
+            // https://github.com/webpack/webpack/blob/master/lib/index.js#L267
+            // https://github.com/webpack/webpack/blob/v4.44.2/lib/webpack.js#L90
+            const ValidationError = webpack.ValidationError ? webpack.ValidationError : webpack.WebpackOptionsValidationError;
+
+            // In case of schema errors print and exit process
+            // For webpack@4 and webpack@5
+            if (error instanceof ValidationError) {
+                logger.error(error.message);
+            } else {
+                logger.error(error);
+            }
+
+            process.exit(2);
+        }
+
+        return compiler;
     }
 
     async getCompiler(args, cliOptions) {
         await this.processArgs(args, cliOptions);
-        await this.compilation.createCompiler(this.compilerConfiguration);
-        return this.compilation.compiler;
+
+        return this.createCompiler(this.compilerConfiguration);
     }
 
     async run(args, cliOptions) {
         await this.processArgs(args, cliOptions);
-        await this.compilation.createCompiler(this.compilerConfiguration);
-        const webpack = await this.compilation.webpackInstance({
-            options: this.compilerConfiguration,
-            outputOptions: this.outputConfiguration,
-        });
-        return webpack;
+
+        const compiler = this.createCompiler(this.compilerConfiguration);
+
+        const options = this.compilerConfiguration;
+        const outputOptions = this.outputConfiguration;
+
+        if (outputOptions.interactive) {
+            const interactive = require('./utils/interactive');
+
+            return interactive(options, outputOptions);
+        }
+
+        const compilers = compiler.compilers ? compiler.compilers : [compiler];
+        const isWatchMode = Boolean(compilers.find((compiler) => compiler.options.watch));
+        const isRawOutput = typeof outputOptions.json === 'undefined';
+
+        if (isRawOutput) {
+            for (const compiler of compilers) {
+                if (outputOptions.progress) {
+                    const { ProgressPlugin } = webpack;
+
+                    let progressPluginExists;
+
+                    if (compiler.options.plugins) {
+                        progressPluginExists = Boolean(compiler.options.plugins.find((e) => e instanceof ProgressPlugin));
+                    }
+
+                    if (!progressPluginExists) {
+                        new ProgressPlugin().apply(compiler);
+                    }
+                }
+            }
+
+            compiler.hooks.watchRun.tap('watchInfo', (compilation) => {
+                if (compilation.options.bail && isWatchMode) {
+                    logger.warn('You are using "bail" with "watch". "bail" will still exit webpack when the first error is found.');
+                }
+
+                logger.success(`Compilation${compilation.name ? `${compilation.name}` : ''} starting...`);
+            });
+            compiler.hooks.done.tap('watchInfo', (compilation) => {
+                logger.success(`Compilation${compilation.name ? `${compilation.name}` : ''} finished`);
+            });
+        }
+
+        const callback = (error, stats) => {
+            if (error) {
+                logger.error(error);
+                process.exit(1);
+            }
+
+            if (stats.hasErrors()) {
+                process.exitCode = 1;
+            }
+
+            const getStatsOptions = (stats) => {
+                // TODO remove after drop webpack@4
+                if (webpack.Stats && webpack.Stats.presetToOptions) {
+                    if (!stats) {
+                        stats = {};
+                    } else if (typeof stats === 'boolean' || typeof stats === 'string') {
+                        stats = webpack.Stats.presetToOptions(stats);
+                    }
+                }
+
+                stats.colors = typeof stats.colors !== 'undefined' ? stats.colors : coloretteOptions.enabled;
+
+                return stats;
+            };
+
+            const foundStats = compiler.compilers
+                ? { children: compiler.compilers.map((compiler) => getStatsOptions(compiler.options.stats)) }
+                : getStatsOptions(compiler.options.stats);
+
+            if (outputOptions.json === true) {
+                process.stdout.write(JSON.stringify(stats.toJson(foundStats), null, 2) + '\n');
+            } else if (typeof outputOptions.json === 'string') {
+                const JSONStats = JSON.stringify(stats.toJson(foundStats), null, 2);
+
+                try {
+                    writeFileSync(outputOptions.json, JSONStats);
+
+                    logger.success(`stats are successfully stored as json to ${outputOptions.json}`);
+                } catch (error) {
+                    logger.error(error);
+
+                    process.exit(2);
+                }
+            } else {
+                logger.raw(`${stats.toString(foundStats)}`);
+            }
+
+            if (isWatchMode) {
+                logger.success('watching files for updates...');
+            }
+        };
+
+        if (isWatchMode) {
+            const watchOptions = (compiler.options && compiler.options.watchOptions) || {};
+
+            if (watchOptions.stdin) {
+                process.stdin.on('end', function () {
+                    process.exit();
+                });
+                process.stdin.resume();
+            }
+
+            return new Promise((resolve) => {
+                compiler.watch(watchOptions, (error, stats) => {
+                    callback(error, stats);
+
+                    resolve();
+                });
+            });
+        } else {
+            return new Promise((resolve) => {
+                compiler.run((error, stats) => {
+                    if (compiler.close) {
+                        compiler.close(() => {
+                            callback(error, stats);
+
+                            resolve();
+                        });
+                    } else {
+                        callback(error, stats);
+
+                        resolve();
+                    }
+                });
+            });
+        }
     }
 }
